+++ src/Map.ts
@@ -7,7 +7,7 @@ import type {
   StyleSwapOptions,
   StyleOptions,
   MapDataEvent,
-  Tile,
+  // Tile,
   RasterDEMSourceSpecification,
   RequestTransformFunction,
   Source,
@@ -17,6 +17,7 @@ import type {
   FilterSpecification,
   StyleSetterOptions,
   ExpressionSpecification,
+  SymbolLayerSpecification,
 } from "maplibre-gl";
 import { ReferenceMapStyle, MapStyleVariant } from "@maptiler/client";
 import { config, MAPTILER_SESSION_ID, SdkConfig } from "./config";
@@ -34,12 +35,13 @@ import { MaptilerTerrainControl } from "./MaptilerTerrainControl";
 import { MaptilerNavigationControl } from "./MaptilerNavigationControl";
 import { geolocation } from "@maptiler/client";
 import { MaptilerGeolocateControl } from "./MaptilerGeolocateControl";
-import { AttributionControl } from "./AttributionControl";
-import { ScaleControl } from "./ScaleControl";
-import { FullscreenControl } from "./FullscreenControl";
+import { AttributionControl } from "./MLAdapters/AttributionControl";
+import { ScaleControl } from "./MLAdapters/ScaleControl";
+import { FullscreenControl } from "./MLAdapters/FullscreenControl";
 
 import Minimap from "./Minimap";
 import type { MinimapOptionsInput } from "./Minimap";
+import { registerLocalCacheProtocol } from "./caching";
 
 export type LoadWithTerrainEvent = {
   type: "loadWithTerrain";
@@ -60,7 +62,7 @@ export const GeolocationType: {
 
 type MapTerrainDataEvent = MapDataEvent & {
   isSourceLoaded: boolean;
-  tile: Tile;
+  // tile: Tile;
   sourceId: string;
   source: RasterDEMSourceSpecification;
 };
@@ -101,6 +103,11 @@ export type MapOptions = Omit<MapOptionsML, "style" | "maplibreLogo"> & {
    * For free plans: MapTiler logo always shows, regardless of the value.
    */
   maptilerLogo?: boolean;
+
+  /**
+   * Attribution text to show in an {@link AttributionControl}.
+   */
+  customAttribution?: string | Array<string>;
 
   /**
    * Enables 3D terrain if `true`. (default: `false`)
@@ -205,6 +212,10 @@ export class Map extends maplibregl.Map {
       maplibreLogo: false,
       transformRequest: combineTransformRequest(options.transformRequest),
     });
+
+    if (config.caching) {
+      registerLocalCacheProtocol();
+    }
 
     this.primaryLanguage = options.language ?? config.primaryLanguage;
     this.forceLanguageUpdate =
@@ -981,7 +992,34 @@ export class Map extends maplibregl.Map {
 
     const { layers } = this.getStyle();
 
-    for (const { id, layout } of layers) {
+    for (const genericLayer of layers) {
+      // Only symbole layer can have a layout with text-field
+      if (genericLayer.type !== "symbol") {
+        continue;
+      }
+
+      const layer = genericLayer as SymbolLayerSpecification;
+      const source = this.getSource(layer.source);
+
+      // Only a layer that is bound to a valid source is considered for language switching
+      if (!source) {
+        continue;
+      }
+
+      // Only source with a url are considered
+      if (!("url" in source && typeof source.url === "string")) {
+        continue;
+      }
+
+      const sourceURL = new URL(source.url);
+
+      // Only layers managed by MapTiler are considered for language switch
+      if (sourceURL.host !== defaults.maptilerApiHost) {
+        continue;
+      }
+
+      const { id, layout } = layer;
+
       if (!layout) {
         continue;
       }
@@ -1325,28 +1363,5 @@ export class Map extends maplibregl.Map {
   ): this {
     super.setTransformRequest(combineTransformRequest(transformRequest));
     return this;
-  }
-
-  /**
-   * Loads an image. This is an async equivalent of `Map.loadImage`
-   */
-  async loadImageAsync(
-    url: string,
-  ): Promise<HTMLImageElement | ImageBitmap | null | undefined> {
-    return new Promise((resolve, reject) => {
-      this.loadImage(
-        url,
-        (
-          error: Error | null | undefined,
-          image: HTMLImageElement | ImageBitmap | null | undefined,
-        ) => {
-          if (error) {
-            reject(error);
-            return;
-          }
-          resolve(image);
-        },
-      );
-    });
   }
 }