+++ CHANGELOG.md
@@ -5,6 +5,17 @@
 ### Bug Fixes
 ### Others
 
+## 2.0.0
+### New Features
+- Updating with MapLibre GL JS v4.1
+- Added client-side caching mechanism for tiles and fonts
+### Bug Fixes
+- Language switching now only occurs on layers fueled by a MapTiler Cloud source
+### Others
+- Removed `Map.loadImageAsync()` as MapLibre's `.loadImage()` is now promise based.
+- Added all the exports that are now exposed from MapLibre
+- Adapted some newly exported classes to make them TypeScript-compatible with the `Map` classes from both MapLibre and MapTiler SDK
+
 ## 1.2.1
 ### New Features
 - Elevation lookup at `maptilersdk.elevation` with the function `.at()`, `.batch()` and for geojson payloads. From the update of the MapTiler Client library
@@ -134,7 +145,7 @@
 ### New Features
 - The `apiKey` can now be specified in the `Map` constructor (will propagate to `config`)
 - The `language` can now be speficifed in the `Map` constructo (will **not** propagete to `config` and will apply only to this specific instance)
-- `Map` now has the method `.getSdkConfig()` to retrieve the config object. 
+- `Map` now has the method `.getSdkConfig()` to retrieve the config object.
 - `Map` now has the method `.getMaptilerSessionId()` to retrieve the MapTiler session ID
 Both `.getSdkConfig()` and `.getMaptilerSessionId()` are handy for layers or control built outside of the SDK that still need some of the configuration to interact with the server. Those components do not always have access to the internal of the SDK (especially that the config is scoped) but can access to the `Map` instance to which they are added with the implementation of the `.onAdd()` method.